GFORMS-3823: PopulateAtl code is now a separate service.
